+++ Mindspace_front_next/app/api/hooks/queries/node.ts
