+++ Mindspace_front_next/app/components/LayoutProvider.tsx
