+++ Mindspace_front_next/app/map/components/WriteModal/index.tsx
@@ -8,7 +8,7 @@ import { WriteModalProps } from "@/constants/types";
 import CustomModal from "@/components/CustomModal";
 import { nodeAtom } from "@/recoil/state/nodeAtom";
 import { useRecoilValue } from "recoil";
-import { useUserPostGetQuery } from "@/hooks/queries/board";
+import { useUserPostGetQuery } from "@/api/hooks/queries/board";
 import WriteEditor from "./components/WriteEditor";
 import ReadViewer from "./components/ReadViewer";
